+++ CHANGELOG.md
@@ -6,6 +6,13 @@ in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.1] - 2026-02-14
+
+### Added
+
+- Terminal window using configured / default Eagle binary directory.
+- Simplified packaging process via use of the bundling feature.
+
 ## [1.0.0] - 2026-02-12
 
 ### Added